# packages/studio/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InkOS Studio API",
    "version": "1.5.0",
    "description": "Auto-generated OpenAPI spec from 157 endpoints across 33 route modules."
  },
  "servers": [
    {
      "url": "http://localhost:4579",
      "description": "InkOS Studio API server"
    }
  ],
  "paths": {
...
    "/api/v1/books/{id}/chapter-goals": {
      "get": {
        "tags": [
          "chapter-intent"
        ],
        "summary": "GET /api/v1/books/:id/chapter-goals",
        "operationId": "get__api_v1_books_id_chapter_goals",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/chapter-goals/{chapterNumber}": {
      "put": {
        "tags": [
          "chapter-intent"
        ],
        "summary": "PUT /api/v1/books/:id/chapter-goals/:chapterNumber",
        "operationId": "put__api_v1_books_id_chapter_goals_chapterNumber",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "chapter-intent"
        ],
        "summary": "DELETE /api/v1/books/:id/chapter-goals/:chapterNumber",
        "operationId": "delete__api_v1_books_id_chapter_goals_chapterNumber",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/chapter-intents": {
      "get": {
        "tags": [
          "chapter-intent"
        ],
        "summary": "GET /api/v1/books/:id/chapter-intents",
        "operationId": "get__api_v1_books_id_chapter_intents",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/chapter-intents/{chapterNumber}": {
      "put": {
        "tags": [
          "chapter-intent"
        ],
        "summary": "PUT /api/v1/books/:id/chapter-intents/:chapterNumber",
        "operationId": "put__api_v1_books_id_chapter_intents_chapterNumber",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "chapter-intent"
        ],
        "summary": "DELETE /api/v1/books/:id/chapter-intents/:chapterNumber",
        "operationId": "delete__api_v1_books_id_chapter_intents_chapterNumber",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/chapter-intents/{chapterNumber}/suggestions": {
      "get": {
        "tags": [
          "chapter-intent"
        ],
        "summary": "GET /api/v1/books/:id/chapter-intents/:chapterNumber/suggestions",
        "operationId": "get__api_v1_books_id_chapter_intents_chapterNumber_suggestions",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/chapters/{chapterNumber}/endpoint-check": {
      "get": {
        "tags": [
          "chapter-intent"
        ],
        "summary": "GET /api/v1/books/:id/chapters/:chapterNumber/endpoint-check",
        "operationId": "get__api_v1_books_id_chapters_chapterNumber_endpoint_check",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/interview": {
      "get": {
        "tags": [
          "chapter-intent"
        ],
        "summary": "GET /api/v1/books/:id/interview",
        "operationId": "get__api_v1_books_id_interview",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/books/{id}/event-chain": {
      "get": {
        "tags": [
          "event-chain"
        ],
        "summary": "GET /api/v1/books/:id/event-chain",
        "operationId": "get__api_v1_books_id_event_chain",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/event-chain/extract": {
      "post": {
        "tags": [
          "event-chain"
        ],
        "summary": "POST /api/v1/books/:id/event-chain/extract",
        "operationId": "post__api_v1_books_id_event_chain_extract",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}